LayoutTests/http/tests/security/contentSecurityPolicy/1.1/plugintypes-url-01-expected.txt

This tests our handling of `data:` URLs, given a `plugin-types` CSP directive. Consider this test passing if each of the following frames contains "PASS" or no text at all, and no console warnings appear above.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS.



--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS.



--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
PASS.



--------
Frame: '<!--framePath //<!--frame3-->-->'
--------
